Additional config to Tailwind, updates to mock data and input fields
