+++ src/app/reports/[id]/page.tsx
@@ -29,14 +29,13 @@ const fetchReportById = async (id: string): Promise<Report> => {
         };
     }
 
-    const res = await axios.get(`http://localhost:8080/api/reports/${id}`);
+    const res = await axios.get(`/api/reports/${id}`);
     return res.data;
 };
 
 export default async function ReportsViewPage(
     { params } : { params: { id: string } }
 ) {
-
     const { 
         statusType, 
         submittedType, 
@@ -49,10 +48,10 @@ export default async function ReportsViewPage(
 
     return (
         <>
-            <form>
+            <form className='relative w-full flex flex-col gap-y-4 min-h-full'>
                 {/* Overview: Status, Submitted By, Incident Type */}
                 <div className='overview w-full'>
-                    <h3 className='font-semibold text-base text-primary'>
+                    <h3 className='font-semibold text-base text-primary mb-2'>
                         Overview
                     </h3>
                     <div className='w-full gap-y-4 flex flex-col'>
@@ -70,7 +69,8 @@ export default async function ReportsViewPage(
                             placeholder='' 
                             label='Submitted Type' 
                             value={submittedType} 
-                            icon="/icons/SPM_shield.svg"
+                            icon="/icons/textsms.svg"
+                            disabled={true}
                             readOnly={true} 
                             width="1/2" />
                         </div>
@@ -84,7 +84,7 @@ export default async function ReportsViewPage(
                 </div>
                 {/* Incident Details: Location, Date-Time, Details */}
                 <div className='incident-details w-full'>
-                    <h3 className='font-semibold text-base text-primary'>Details</h3>
+                    <h3 className='font-semibold text-base text-primary mb-2'>Details</h3>
                     <div className='w-full gap-y-4 flex flex-col'>
                         <InputField name='location'
                             placeholder=''
@@ -114,9 +114,14 @@ export default async function ReportsViewPage(
                         value={details}
                         icon='/icons/description.svg'
                         readOnly={true}
-                        disabled={statusType === 'in-progress' ? false : true}
+                        disabled={statusType === 'in progress' ? false : true}
                         />
                     </div>
+                </div>
+                <div className='fixed p-8 left-0 bottom-0 w-full flex flex-col items-center justify-center'>
+                    <button className={`${statusType === 'unclaimed' ? 'bg-primary text-white' : 'bg-white' } rounded-full p-4 shadow-[0px_4px_8px_3px_rgba(0,0,0,0.15),0px_1px_3px_0px_rgba(0,0,0,0.30)] w-full text-center font-semibold`}>
+                        {statusType === 'unclaimed' ? 'Claim Report' : 'Close Report'}
+                    </button>
                 </div>
             </form>
         </>
