+++ src/app/reports/layout.tsx
@@ -8,7 +8,7 @@ import React from 'react'
 export default function ReportsViewLayout({ children } : { children : React.ReactNode }) {
 
     return (
-        <div className='p-8'>
+        <div className='p-8 min-h-screen flex flex-col gap-4'>
             {children}
         </div>
     )
